Extracting models and view into sections
