+++ ios/brave-ios/Sources/BraveVPN/BraveVPN.swift
@@ -585,48 +585,6 @@ public class BraveVPN {
     Task { @MainActor in
       allCountryRegions = await BraveVPNRegionManager.shared.getAllRegionsWithDetails()
     }
-
-    // TODO: Remove Debugging information
-
-    //    helper.setPreferredRegionPrecision(kGRDRegionPrecisionCityByCountry)
-    //
-    //    print("Precision \(helper.regionPrecision)")
-    //
-    //    let delayTime = DispatchTime.now() + .seconds(2)
-    //    DispatchQueue.main.asyncAfter(deadline: delayTime) {
-    //      helper.allRegions { regions, error in
-    //        print("All Regions Default with HELPER")
-    //
-    //        print("=======================")
-    //
-    //        print("\(regions)")
-    //
-    //        print("=======================")
-    //
-    //      }
-    //
-    //      let serverManagerx = GRDServerManager(
-    //        regionPrecision: helper.regionPrecision,
-    //        serverFeatureEnvironment: helper.featureEnvironment,
-    //        betaCapableServers: false
-    //      )
-    //
-    //      serverManagerx.allRegions { regions, error in
-    //
-    //        print("All Regions Default with MANAGER")
-    //
-    //        print("=======================")
-    //
-    //        print("\(regions)")
-    //
-    //        print("=======================")
-    //
-    //      }
-    //
-    //      Task { @MainActor in
-    //        await BraveVPNRegion.shared.getAllRegionsWithDetails()
-    //      }
-    //    }
   }
 
   // MARK: - VPN Alerts and notifications
